requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
schedule>=1.2.0
Pillow>=10.0.0
moviepy>=1.0.3
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
facebook-sdk>=3.1.0
python-telegram-bot>=20.0
colorama>=0.4.6
rich>=13.0.0
pydantic>=2.0.0
pyyaml>=6.0.0
pytz>=2023.3
